rag_system/requirements.txt
# RAG System Dependencies - Groq + Local Embeddings Edition
# Architecture:
#   - Embeddings: sentence-transformers (local)
#   - Generation: Groq API (OpenAI-compatible)
#   - Vector DB: FAISS

# =============================================================================
# PDF Processing
# =============================================================================
pdfplumber>=0.10.0

# =============================================================================
# Local Embeddings (sentence-transformers)
# =============================================================================
sentence-transformers>=2.2.0
torch>=2.0.0

# =============================================================================
# Vector Search
# =============================================================================
faiss-cpu>=1.7.0

# =============================================================================
# Groq API (OpenAI-compatible client)
# =============================================================================
openai>=1.0.0
python-dotenv>=1.0.0

# =============================================================================
# API Server
# =============================================================================
fastapi>=0.108.0
uvicorn>=0.25.0
python-multipart>=0.0.6

# =============================================================================
# Utilities
# =============================================================================
numpy>=1.24.0
tqdm>=4.66.0
pydantic>=2.5.0
